+++ Assets/Scripts/SGG.PerfMeter/Tests/EditMode/PerformanceMeterApiTests.cs
@@ -62,7 +62,7 @@ namespace SGG.PerfMeter.Tests.EditMode
 			Assert.That(status.GraphicsDeviceName, Is.Not.Null);
 			Assert.That(status.OverlayMode, Is.EqualTo(PerfMeterOverlayMode.Full));
 			Assert.That(status.OverlayLayout, Is.EqualTo(PerfMeterOverlayLayout.MetricBars));
-			AssertHasModule(status.OverlayModules, PerfMeterOverlayModule.CpuCoreBars);
+			AssertDoesNotHaveModule(status.OverlayModules, PerfMeterOverlayModule.CpuCoreBars);
 			AssertDoesNotHaveModule(status.OverlayModules, PerfMeterOverlayModule.CpuCores);
 			AssertDoesNotHaveModule(status.OverlayModules, PerfMeterOverlayModule.CpuCoreGraphs);
 			Assert.That(status.TargetFps, Is.EqualTo(PerfMeterTargetFps.Fps60));
@@ -91,7 +91,7 @@ namespace SGG.PerfMeter.Tests.EditMode
 			Assert.That(status.OverlayLayout, Is.EqualTo(PerfMeterOverlayLayout.MetricBars));
 			Assert.That(status.OverlayModules, Is.Not.EqualTo(PerfMeterOverlayModule.All));
 			AssertHasModule(status.OverlayModules, PerfMeterOverlayModule.Graphs);
-			AssertHasModule(status.OverlayModules, PerfMeterOverlayModule.CpuCoreBars);
+			AssertDoesNotHaveModule(status.OverlayModules, PerfMeterOverlayModule.CpuCoreBars);
 			AssertDoesNotHaveModule(status.OverlayModules, PerfMeterOverlayModule.CpuCores);
 			AssertDoesNotHaveModule(status.OverlayModules, PerfMeterOverlayModule.CpuCoreGraphs);
 			Assert.That(status.SessionState, Is.EqualTo(PerfMeterSessionState.Idle));
@@ -176,7 +176,7 @@ namespace SGG.PerfMeter.Tests.EditMode
 			Assert.That(PerformanceMeter.OverlayCorner, Is.EqualTo(PerfMeterOverlayCorner.TopRight));
 			Assert.That(PerformanceMeter.OverlayMode, Is.EqualTo(PerfMeterOverlayMode.Full));
 			Assert.That(PerformanceMeter.OverlayLayout, Is.EqualTo(PerfMeterOverlayLayout.MetricBars));
-			AssertHasModule(PerformanceMeter.OverlayModules, PerfMeterOverlayModule.CpuCoreBars);
+			AssertDoesNotHaveModule(PerformanceMeter.OverlayModules, PerfMeterOverlayModule.CpuCoreBars);
 			AssertDoesNotHaveModule(PerformanceMeter.OverlayModules, PerfMeterOverlayModule.CpuCores);
 			AssertDoesNotHaveModule(PerformanceMeter.OverlayModules, PerfMeterOverlayModule.CpuCoreGraphs);
 			Assert.That(PerformanceMeter.TargetFps, Is.EqualTo(PerfMeterTargetFps.Fps60));